tests/wpt/tests/html/semantics/embedded-content/the-iframe-element/iframe_javascript_url_form_submission.html

<!DOCTYPE html>
<title>javascript: URL in form submission, initial src is not about:blank</title>
<link rel="help" href="https://html.spec.whatwg.org/multipage/#process-the-iframe-attributes">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>
<script>
setup({single_test: true});
let iframeLoaded = false;
window.javascriptUrlRan = false;
</script>
<iframe src="resources/javascript_url_form_submission.html" onload="iframeLoaded = true; this.onload = assert_unreached;"></iframe>
<script>
onload = () => {
  const iframe = document.querySelector('iframe');
  const btn = iframe.contentDocument.querySelector("button");
  btn.onclick = (e) => {
    const img = iframe.contentDocument.querySelector("img");
    const newImg = iframe.contentDocument.createElement("img");
    iframe.contentDocument.documentElement.appendChild(newImg);
    newImg.src = img.src;
  };

  assert_true(iframeLoaded, "iframeLoaded");
  btn.click();
  assert_false(javascriptUrlRan, "javascriptUrlRan");
  step_timeout(() => {
    assert_true(javascriptUrlRan, "javascriptUrlRan");
    done();
  }, 100); // Verify only one load event is fired on iframe
};
</script>
